+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2021-11-30
+
 ### Added
 - sort queried events (#417)
 - expose basic metrics from server, chaincode and forwarder behind `METRICS_ENABLED` feature flag